Events are being added to the database with the information on the start date and end date.
Changed events table in the DB file and corrected code in the DataSource class to match the current structure. Updated methods in the Events class to match the information that needs to be passed and retrieved from DB. Renamed fields in the Event class to match the current code structure.