Working on migrating from JDBC Framework to Spring Data JPA - created Repository interfaces for Plant and RecurringEvent, fixed minor bugs, rearranged DB
Fixed a bug in the getAllEventsInTheDateRange() method from RecurringEvent class. Now the method is correctly calculating the number of days between the startDate and from date.
Added @Enumerated in Plant to the type field.
Added @EnableJpaRepositiories and @EntityScan in the PlantWateringApplication class.
Removed the aliases table from the DB updated queries and removed redundant code.