Working on migrating from JDBC Framework to Spring Data JPA - corrected method to retrieve all results in the date range in RecurringEvent; fixed bugs connected to timestamp parsing error in RecurringEvent, corrected Plant and RecurringEvent to reflect the tables in the DB. Changed SQLite to Postgres.
Corrected getAllEventsInTheDateRange(LocalDate from, LocalDate to) in RecurringEvent to retrieve the correct amount of results when the number of days between the start date of the recurring event and the beginning of the range provided by the user exceeds one month. Added test to check for this issue.
Changed the startDate and EndDate in RecurringEvent from LocalDate to LocalDateTime, to match the SQLite requirements and corrected the getter and setter methods as well as methods in MyDataSource to reflect this change.
Fixed MyPlantsController to display the amount of plants available in the DB rather than set amount.
Corrected Plant and RecurringEvent to reflect the tables in the DB, set correct fields to not null.
Changed application.properties to use Postgres instead of SQLite.