Updated the functionality for saving plants in the DB.
Updated MyDataSource and MyPlantsController and MyPlantsService to allow the plants to be saved correctly in the DB.
Added options to the dropdown menu for type of plant. Now user has the option to choose from any type available in the PlantType enum.
Corrected a bug that prevented the validation from being displayed.
Added validation, to prevent duplicate plants from being added to the database.