Updated the functionality for saving events in the DB.
Updated MyDataSource and MyPlantsController and EventsService to allow the events to be saved correctly in the DB.
Added options to the dropdown menu for type of event. Now user has the option to choose from any type available in the EventType enum.
Added validation, to prevent duplicate events from being added to the database.
Corrected logic in the IJpaRecurringEventRepository to allow the program to handle events without specified end date and added tests for this functionality.
Added a form on the MyPlants page to allow users to type in the details of the event.